tests: lib: logging: move o_helper here
This needs to be imported the same way for isinstance() to work on this
object so let's move it to a library py

Reviewed-on: https://review.gerrithub.io/c/lgirdk/boardfarm/+/449268
