+++ tests/lib/logging.py
@@ -9,6 +9,8 @@ import time
 import types
 from datetime import datetime
 from devices.common import print_bold
+from termcolor import colored
+import re
 
 def now_short(_format = "%Y%m%d-%H%M%S"):
     """
@@ -62,3 +64,30 @@ def log_message(s, msg, header = False):
     else:
         print_bold(full_msg)
         s.log_to_file += now_short()+msg+"\r\n"
+
+class o_helper(object):
+    def __init__(self, parent, out, color):
+        self.color = color
+        self.out = out
+        self.parent = parent
+        self.first_write = True
+    def write(self, string):
+        if self.first_write:
+            self.first_write = False
+            string = "\r\n" + string
+        if self.color is not None:
+            self.out.write(colored(string, self.color))
+        else:
+            self.out.write(string)
+        td = datetime.now()-self.parent.start
+        # check for the split case
+        if len(self.parent.log) > 1 and self.parent.log[-1] == '\r' and string[0] == '\n':
+            tmp = '\n[%s]' % td.total_seconds()
+            tmp += string[1:]
+            string = tmp
+        to_log = re.sub('\r\n', '\r\n[%s]' % td.total_seconds(), string)
+        self.parent.log += to_log
+        if hasattr(self.parent, 'test_to_log'):
+            self.parent.test_to_log.log += "%s: " % re.sub('\r\n\[', '\r\n%s: [' % self.parent.test_prefix, to_log)
+    def flush(self):
+        self.out.flush()
